requirements.txt
tensorflow==2.8.3
tensorboard==1.13.1
tensorflow_hub==0.14.0
flask==2.2.5
werkzeug
flask-cors==4.0.0
matplotlib==3.5.3
ipdb==0.13.13
cython==3.0.0
absl_py==1.4.0
Pillow==8.4.0
wheel==0.41.1
cython==3.0.0
chumpy==0.70
numpy==1.16.4
scipy==1.1.0
#spicy==0.16.0
openpyxl==2.6.2
ecos==2.0.7.post1
vtk==8.1.2
traits==5.1.1
fancyimpute==0.3.2
mayavi==4.7.1
pyopengl==3.1.0
pyqt5==5.10.1
scikit-learn==0.21.3
scikit-image==0.14.2
opencv-python==4.2.0.32
six
#opendr==0.73


#cvxpy==1.0.21
